- Extract private methods to a service
- Some minor refactoring
