+++ AngularLibraryPackageJsonSync/Models/AngularJson.cs
@@ -5,5 +5,10 @@ namespace AngularLibraryPackageJsonSync.Models;
 public class AngularJson
 {
     [JsonPropertyName("projects")]
-    public Dictionary<string, AngularProject> AngularProjects { get; set; }
+    public required Dictionary<string, AngularProject> AngularProjects { get; set; }
+    
+    public IEnumerable<KeyValuePair<string, AngularProject>> GetMatchingLibraryProjects(string angularLibraryType)
+    {
+        return AngularProjects.Where((p) => string.Equals(p.Value.ProjectType, angularLibraryType, StringComparison.CurrentCultureIgnoreCase));
+    }
 }
